Consolidated gyro code, added RobotMap comments
