+++ src/main/java/frc/robot/RobotMap.java
@@ -13,24 +13,24 @@ public class RobotMap {
   public static final int powerDistributionPanel = 0;
 
   public static final int leftMotor1 = 10;
-  public static final int leftMotor2 = 11;       // Talon
+  public static final int leftMotor2 = 11;       // Talon with left drive encoder
   public static final int leftMotor3 = 12;
   public static final int rightMotor1 = 20;
-  public static final int rightMotor2 = 21;      // Talon
+  public static final int rightMotor2 = 21;      // Talon with right drive encoder
   public static final int rightMotor3 = 22;
   
-  public static final int elevatorMotor1 = 30;   // Talon
-  public static final int elevatorMotor2 = 31;   // Talon
+  public static final int elevatorMotor1 = 30;   // Talon with elevator encoder and UL/LL switches
+  public static final int elevatorMotor2 = 31;   // Talon (with left drive encoder in Spark config)
 
   public static final int cargoMotor1 = 40;  // Top cargo motor
   public static final int cargoMotor2 = 41;  // Bottom cargo motor
 
-  public static final int climbMotor1 = 50;     // Talon
-  public static final int climbMotor2 = 51;     // Talon
+  public static final int climbMotor1 = 50;     // Talon (with right drive encoder in Spark config)
+  public static final int climbMotor2 = 51;     //  Talon with climb encoder and UL switch
   public static final int climbVacuum1 = 52;
   //public static final int climbVacuum2 = 53;
 
-  public static final int wristMotor = 60;      // Talon
+  public static final int wristMotor = 60;      // Talon with wrist encoder and UL/LL switches
 
   // Pneumatic Addresses
   //public static final int pnuematicShifterLow = 0;
